requirements.txt
django_filter==2.4.0
googletrans==3.1.0a0
Django==3.1.3
django_recaptcha==2.0.6
Pillow==8.3.2
protobuf==3.18.0
django-environ==0.7.0
psycopg2-binary==2.9.1
google-cloud-secret-manager==2.7.2
PyMySQL==0.9.3
google-cloud-vision==1.0.0
gunicorn==19.7.1
